refactor: separate by landing & player
